data_frame func moved to DataHandler class, moved around chunks of code for aesthetic reasons.
