+++ main.py
@@ -32,125 +32,12 @@ import math
 """
 
 
-
-
-def data_frame(
-        data_handler: data_downl.DataHandler,
-        frame_n: int,
-        local_cache: tuple,
-        flow_vectors: np.array = None,
-        flow_lines_visual: np.array = None):
-    n, labels, stats, positions, image1 = local_cache
-
-    image2 = cv2.imread(f'{data_handler.directory}{data_handler.filename}({frame_n + 1}).png')
-    image2 = data_handler.rechannel_image(image2)
-    n2, labels2, stats2, positions2 = watershed.watershed(image2)
-
-    imgy, imgx, _ = image2.shape
-
-    cool_img_visual = image2.copy()  # TEMP!!!!!!
-
-    """ We skip index = 0, because the first object is always the background """
-    for index in range(1, n):
-        x, y, w, h, A = stats[index]
-
-        """ We make the bounding box larger by AT LEAST 1 pixel, to ensure that we
-        are able to detect the objects edges (the bounding box is tightly touching the object otherwise) """
-        object_box_margin = 5
-        if x > 0:
-            x -= object_box_margin
-            if x + w < imgx:
-                w += object_box_margin
-        elif w < imgx:
-            w += object_box_margin
-        if y > 0:
-            y -= object_box_margin
-            if y + h < imgy:
-                h += object_box_margin
-        elif h < imgy:
-            h += object_box_margin
-        # We can rewrite the ^above^ code using min and max functions if needed
-
-        # (s like search):
-        """ We create a bounding box with a few rules to avoid IndexError: """
-        sy, sx, sh, sw = data_analyser.create_bounds(image2, (y, x), (h, w))
-
-        search_area = image2[sy:sy + sh, sx:sx + sw]
-
-        """ Instead of image there was output before, but idk why cuz this works far better: """
-        search_data = image1.copy()
-        search_data = search_data[y:y + h, x:x + w]
-
-        # (f like found):
-        """ We search for the first image inside the second one: """
-        fy, fx, fh, fw = data_analyser.find_element_in(search_data, search_area)
-
-        if (fy, fx) == (0, 0):
-            """ The object was not found in the image, therefore saving it would only create noise. """
-            continue
-
-        # print(index, f'v=[{sx-x+fx} {sy-y+fy}], f=[{fx} {fy}]')
-
-        a = (x, y)
-        b = (x + w, y + h)
-        cv2.rectangle(cool_img_visual, a, b, (0, 255, 255), 1)
-        a = (sx + fx, sy + fy)
-        b = (sx + fx + fw, sy + fy + fh)
-        cv2.rectangle(cool_img_visual, a, b, (0, 0, 255), 1)
-
-        cv2.putText(
-            cool_img_visual,
-            f'{index}',
-            (a[0], a[1] - 5),
-            cv2.FONT_HERSHEY_SIMPLEX,
-            0.3,
-            (0, 0, 255),
-            1,
-            cv2.LINE_AA
-        )
-
-        old_center = (x + w // 2, y + h // 2)
-        new_center = (sx + fx + fw // 2, sy + fy + fh // 2)
-
-        cv2.line(cool_img_visual, old_center, new_center, (0, 255, 0), 1)
-        cv2.imshow('Search result:', cool_img_visual)
-
-        cv2.line(flow_lines_visual, old_center, new_center, (0, 255, 0), 1)
-        cv2.imshow('flowlines(nested)', flow_lines_visual)
-
-        # v = [ x, y, dx, dy, A2/A[%.] ]
-        """ We find the object closest to our objective's  location """
-        distances = data_analyser.arr_elements_distances(
-            stats2[:, :2],
-            np.asarray([x, y])
-        )
-        min_dist_index = np.argmin(distances)
-
-        dA = round(stats2[min_dist_index][4] * 1000 / A)  # area change in promiles
-        # print(sx - x + fx, sy - y + fy, dA, )
-        flow_vectors.append((old_center[0], old_center[1], sx - x + fx, sy - y + fy, dA))
-        # flow_vectors[old_center] = (32767 + sx - x + fx, 32767 + sy - y + fy, dA)
-
-        # we add (2^16-1)//2 to avoid
-        # "negative overflow"
-        # OUTDATED!!! now we save as a float array either way, so we don't have to
-        # worry about any of that.
-
-        search_area_viewport = np.zeros((640, 640, 3), np.uint8)
-        search_area_viewport[sy:sy + sh, sx:sx + sw] = image1[sy:sy + sh, sx:sx + sw]
-        cv2.rectangle(search_area_viewport, (sx, sy), (sx + sw, sy + sh), (0, 255, 0), 1)
-    cv2.waitKey(10)
-    """ Saving the objects' data in place of the raw data images. (Debugging's sake) """
-    cv2.imwrite(f'{data_handler.directory}{data_handler.filename}({frame_n}).png', cool_img_visual)
-    return n2, labels2, stats2, positions2, image2
-
-
 def main():
     handler = data_downl.DataHandler(directory='data_directory')
     handler.update_name()  # creates name
     handler.download_data()  # downloads data
 
-    # generating a duplicate of the last frame
+    # generates a duplicate of the last frame:
     last_frame = cv2.imread(
         f'{handler.directory}{handler.filename}({handler.n_frames-1}).png'
     )
@@ -158,16 +45,17 @@ def main():
 
     handler.convert_frames()  # extracts data
 
+    """ Main data extraction loop: """
+
+    # initiating data before the loop
     image0 = cv2.imread(f'{handler.directory}{handler.filename}(0).png')
     image0 = handler.rechannel_image(image0)
     local_cache = watershed.watershed(image0) + (image0,)
-
-    """ Main data extraction loop: """
     flow_lines_visual = np.zeros(handler.image.shape + (3,), np.uint8)
     flow_vector_arr = []
+
     for frame_n in range(handler.n_frames - 1):  # handler.n_frames-1
-        local_cache = data_frame(
-            data_handler=handler,
+        local_cache = handler.data_frame(
             frame_n=frame_n,
             local_cache=local_cache,
             flow_vectors=flow_vector_arr,
@@ -176,6 +64,7 @@ def main():
     flow_vector_arr = np.asarray(flow_vector_arr)
 
     """ Removing null vectors: """
+
     print(f'Removing null vectors...', end='\r')
     removed_count = 0
     for index, vector in enumerate(flow_vector_arr):
@@ -186,6 +75,7 @@ def main():
 
     """ Adding a general direction to all vectors and dividing by 2 (to turn all of them kinda in
         the same direction) """
+
     flow_vector_arr_mean = np.average(flow_vector_arr, axis=0)
     flow_vector_arr_mean[:2] = 0
     flow_vector_arr_mean[4:] = 0
@@ -195,10 +85,12 @@ def main():
     flow_vector_arr = np.round(flow_vector_arr).astype(np.int32)
 
     """ Saving the flow_lines array for debugging's sake: """
+
     cv2.imwrite(f'{handler.directory}{handler.filename}[main_vectors].png', flow_lines_visual)
 
 
     """ Saving the initial vector array: """
+
     with open(f'{handler.directory}{handler.filename}[main_vectors].npy', 'wb') as f:
         np.save(f, flow_vector_arr)
